Refactor unit tests to remove unnecessary inclusion of MeadeResponse.hpp
- Removed #include "core/MeadeResponse.hpp" from multiple unit test files.
- Updated includes to only reference "core/MeadeParser.hpp" where applicable.
