+++ unit_tests/test_meade_extra/test_MeadeExtra.cpp
@@ -13,7 +13,7 @@
 #include <unity.h>
 
 #include "core/MeadeParser.hpp"
-#include "core/MeadeResponse.hpp"
+#include "core/MeadeParser.hpp"
 
 namespace meade = oat::core::meade;
