+++ unit_tests/test_meade_gps/test_MeadeGps.cpp
@@ -9,7 +9,7 @@
 #include <unity.h>
 
 #include "core/MeadeParser.hpp"
-#include "core/MeadeResponse.hpp"
+#include "core/MeadeParser.hpp"
 
 namespace meade = oat::core::meade;
