+++ unit_tests/test_meade_home/test_MeadeHome.cpp
@@ -7,7 +7,7 @@
 #include <unity.h>
 
 #include "core/MeadeParser.hpp"
-#include "core/MeadeResponse.hpp"
+#include "core/MeadeParser.hpp"
 
 namespace meade = oat::core::meade;
