+++ unit_tests/test_meade_init/test_MeadeInit.cpp
@@ -5,7 +5,7 @@
 #include <unity.h>
 
 #include "core/MeadeParser.hpp"
-#include "core/MeadeResponse.hpp"
+#include "core/MeadeParser.hpp"
 
 namespace meade = oat::core::meade;
