+++ unit_tests/test_meade_movement/test_MeadeMovement.cpp
@@ -7,7 +7,7 @@
 #include <unity.h>
 
 #include "core/MeadeParser.hpp"
-#include "core/MeadeResponse.hpp"
+#include "core/MeadeParser.hpp"
 
 namespace meade = oat::core::meade;
