+++ unit_tests/test_meade_quit/test_MeadeQuit.cpp
@@ -9,7 +9,7 @@
 #include <string.h>
 
 #include "core/MeadeParser.hpp"
-#include "core/MeadeResponse.hpp"
+#include "core/MeadeParser.hpp"
 
 namespace meade = oat::core::meade;
