+++ unit_tests/test_meade_set/test_MeadeSet.cpp
@@ -11,7 +11,7 @@
 #include <string.h>
 
 #include "core/MeadeParser.hpp"
-#include "core/MeadeResponse.hpp"
+#include "core/MeadeParser.hpp"
 
 namespace meade = oat::core::meade;
