+++ unit_tests/test_meade_sync/test_MeadeSync.cpp
@@ -6,7 +6,7 @@
 #include <unity.h>
 
 #include "core/MeadeParser.hpp"
-#include "core/MeadeResponse.hpp"
+#include "core/MeadeParser.hpp"
 
 namespace meade = oat::core::meade;
